requirements.txt
PyPDF2 
langchain 
google-generativeai 
chromadb